requirements.txt
tensorflow-cpu==2.16.1
# TF 2.16 wheels on py310 expect NumPy < 2.0
numpy>=1.24.0,<2.0
pandas==2.2.3
pillow==10.4.0
matplotlib==3.9.2
scikit-learn==1.5.2
opencv-python-headless==4.10.0.84
tqdm==4.66.5

